Add system-setup scripts
This commit adds the system-setup scripts which help with rapidly
building my desktop environment from a fresh Arch install. It consists
of two phases:

Phase 1: Using bash, we create a normal user account, update the system,
and add the new user to the wheel group as well as give the wheel group
sudo privledges. This script is run from the superuser account since
that is the only account present after a fresh Arch install. This phase
also installs fish from the official Arch repo's since phase two is
written as a fish script.

Phase 2: We now logoff of the superuser account and login to our new
user account. Since we now have fish installed, this phase runs as a
fish script. It takes care of installing all the official and unofficial
packages I typically want, including my wayland compositor, texteditor,
latex, dev tools, better CLI utilities, Rustup, bluetooth and GPU
drivers, etc. It also generates a new SSH key for my github to use, and
once I have added it to my profile, the script continues to clone my
dotfiles and link them to the proper places on the system.